+++ BambooTracker/command/commands.hpp
@@ -32,6 +32,7 @@
 #include "./pattern/reverse_pattern_command.hpp"
 #include "./pattern/replace_instrument_in_pattern_command.hpp"
 #include "./pattern/paste_overwrite_copied_data_to_pattern_command.hpp"
+#include "./pattern/change_values_in_pattern_command.hpp"
 
 /********** Order edit **********/
 #include "./order/set_pattern_to_order_command.hpp"