+++ BambooTracker/gui/command/pattern/pattern_commands_qt.hpp
@@ -26,5 +26,6 @@
 #include "reverse_pattern_qt_command.hpp"
 #include "replace_instrument_in_pattern_qt_command.hpp"
 #include "paste_overwrite_copied_data_to_pattern_qt_command.hpp"
+#include "change_values_in_pattern_qt_command.hpp"
 
 #endif // PATTERN_COMMANDS_QT_HPP